Added search bar history functionality
When the user manually types in a city name and searches for it, it is now added to a history list stored in localStorage. The left sidebar list items are now clickable for the user to quickly search for past cities. The history does not update, purposefully, when the user clicks on past search results.